requirements.txt
torch>=2.0.0
numpy>=1.24.0
einops>=0.6.0
pyyaml>=6.0
scikit-learn>=1.2.0
